requirements.txt
torchvision==0.5.0
opencv_python==4.1.2.30
torch==1.4.0
numpy==1.18.1
Pillow==7.1.1
